+++ EssentialFeedTests/Feed API/URLSessionHTTPClientTests.swift
@@ -124,19 +124,10 @@ class URLSessionHTTPClientTests: XCTestCase {
         return recivedValues
     }
     
-    
-    private func anyURL() -> URL {
-        return URL(string: "http://any-url.com")!
-    }
-    
     private func anyData() -> Data {
         return Data("any data".utf8)
     }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 0)
-    }
-    
+        
     private func nonHTTPURLResponse() -> URLResponse {
         return URLResponse(url: anyURL(), mimeType: nil, expectedContentLength: 0, textEncodingName: nil)
     }
