+++ EssentialFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -129,23 +129,5 @@ class CacheFeedUseCaseTests: XCTestCase {
         wait(for: [exp], timeout: 1.0)
         
         XCTAssertEqual(recivedError as NSError?, expectedError, file: file, line: line)
-    }
-        
-    private func uniqueImage() -> FeedImage {
-        FeedImage(id: UUID(), description: "Any", location: "Any", url: anyURL())
-    }
-    
-    private func uniqueImageFeed() -> (model: [FeedImage], local: [LocalFeedImage]) {
-        let model = [uniqueImage(), uniqueImage()]
-        let local = model.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.url) }
-        return (model, local)
-    }
-    
-    private func anyURL() -> URL {
-        return URL(string: "http://any-url.com")!
-    }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 0)
-    }
+    }        
 }
