+++ EssentialFeedTests/Helpers/SharedTestHelpers.swift
@@ -1,0 +1,17 @@
+//
+//  SharedTestHelpers.swift
+//  EssentialFeedTests
+//
+//  Created by Mauro Worobiej on 26/09/2021.
+//
+
+import Foundation
+
+func anyNSError() -> NSError {
+    return NSError(domain: "any error", code: 0)
+}
+
+func anyURL() -> URL {
+    return URL(string: "http://any-url.com")!
+}
+
